work still to do on connecting the new info to the state
