+++ packages/studio/src/app/app.tsx
@@ -1,10 +1,10 @@
-import React from "react";
+import React, { useContext } from "react";
 import { Page } from "@patternfly/react-core";
 import AppHeader from "./appHeader";
 import {BrowserRouter as Router, Route} from 'react-router-dom';
 import * as Pages from './pages';
 import './app.css';
-import {StoreContext} from './../context/StoreContext';
+import { StoreContext } from './../context/StoreContext';
 
 export const App: React.FunctionComponent = () => {
   const { state, dispatch, actions } = useContext(StoreContext);
